build: store hash of quacs-data and use that for rebuilds
Sometimes deployments from scraping fail - if that semester doesnt change on the next update, the update will be delayed indefinitely (either until a new change comes in which rebuilds all the semesters or a new update comes along).
To address this, save the hash of quacs-data in the site repository and use that for partial rebuilds. This way no updates are missed.

This also addresses a race condition when a mass-rebuild occurs and quacs-data scraping is running.